+++ package/dotnetcore/dotnetruntime/modified/apibridge.h
@@ -50,5 +50,5 @@ int local_X509_set1_notBefore(X509* x509, const ASN1_TIME*);
 int local_X509_set1_notAfter(X509* x509, const ASN1_TIME*);
 const X509_ALGOR* local_X509_get0_tbs_sigalg(const X509* x509);
 X509_PUBKEY* local_X509_get_X509_PUBKEY(const X509* x509);
-int32_t local_X509_get_version(const X509* x509);
+long local_X509_get_version(const X509* x509);
 int32_t local_X509_up_ref(X509* x509);
